requirements.txt
Django==4.2.8
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
liac-arff==2.5.0
